fix(COMBAT): 修复战斗伤害系统 ushort 截断 - 完成任务 #2
问题：GetHitStruckDamage 返回 ushort，物理伤害 >65535 时截断
根因：移植错误，原生 sub_767958/sub_7679B8 孪生函数都返回 int32

修复：
- TBaseObject.cs:5284 返回类型 ushort → int
- TBaseObject.cs:5310 移除强制转换
- TBaseObject.Attack.cs:550 移除冗余转换
- 更新 2 个审计工具的断言

影响：65536+ 伤害现在正确显示（之前回环到 0）

验证：
- 构建 0 errors ✓
- 4 个审计工具全部 PASS ✓
- 所有调用点兼容 ✓

状态：任务 #2（战斗伤害系统）✅ 完全闭合
来源：combat-damage-two-pipelines-divergence.md

+++ AuditTools/NativeSkill153ShieldCompatCheck/Program.cs
@@ -203,11 +203,11 @@ static void VerifyPhysicalResolverConsumption()
     // The charge is consumed in StruckDamage (sub_73F9FC @0x73FB5F), NOT in
     // the armour getter: sub_767958 ends `call sub_76FFE8` + `ret 4`
     // (@0x7679A9-0x7679B2) and never reads word [+0x3FC].
-    Equal((ushort)0, actor.GetHitStruckDamage(null, 0),
+    Equal(0, actor.GetHitStruckDamage(null, 0),
         "zero physical damage");
     Equal((ushort)2, Charges(actor),
         "armour getter must not consume a charge");
-    Equal((ushort)200, actor.GetHitStruckDamage(null, 200),
+    Equal(200, actor.GetHitStruckDamage(null, 200),
         "armour getter must pass damage through unshielded");
     Equal((ushort)2, Charges(actor),
         "armour getter consumed a charge on a positive hit");
